backend/requirements.txt
Django>=4.1,<4.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.2
djongo>=1.3.6
pymongo>=4.3
